refactor(rust): group provider routing helpers
