+++ litellm-rust/crates/core/src/lib.rs
@@ -1,8 +1,8 @@
 pub mod error;
-pub mod get_llm_provider_logic;
 pub mod ocr;
 pub mod providers;
 pub mod realtime;
 pub mod router;
+pub mod routing_utils;
 
 pub use error::{CoreError, CoreResult};
